requirements.txt
Pillow
click
